Changing window resolution to 1920x1080 and adapting background to it. Chagging build and run script to check first if build directory exits, before deleting.

+++ SFMLApp/build_and_run.ps1
@@ -1,7 +1,11 @@
-# Remove the "build" directory if it exists
-Write-Host "Removing 'build' directory if it exists..."
-Remove-Item -Path "build" -Recurse -Force
-Write-Host "'build' directory removed if it existed."
+# Check if the "build" directory exists
+if (Test-Path -Path "build") {
+    Write-Host "Directory 'build' exists. Removing it..."
+    Remove-Item -Path "build" -Recurse -Force
+    Write-Host "'build' directory removed."
+} else {
+    Write-Host "'build' directory does not exist. No action needed."
+}
 
 # Get the current working directory (equivalent to pwd in Bash)
 $currentDir = Get-Location
